+++ src/marqo/tensor_search/api.py
@@ -15,6 +15,7 @@ from marqo.api import exceptions as api_exceptions
 from marqo.api.exceptions import InvalidArgError
 from marqo.api.models.embed_request import EmbedRequest
 from marqo.api.models.health_response import HealthResponse
+from marqo.api.models.recommend_query import RecommendQuery
 from marqo.api.models.rollback_request import RollbackRequest
 from marqo.api.models.update_documents import UpdateDocumentsBodyParams
 from marqo.api.route import MarqoCustomRoute
@@ -24,7 +25,7 @@ from marqo.logging import get_logger
 from marqo.tensor_search import tensor_search, utils
 from marqo.tensor_search.enums import RequestType, EnvVars
 from marqo.tensor_search.models.add_docs_objects import (AddDocsBodyParams)
-from marqo.tensor_search.models.api_models import SearchQuery, RecommendQuery
+from marqo.tensor_search.models.api_models import SearchQuery
 from marqo.tensor_search.models.index_settings import IndexSettings, IndexSettingsWithName
 from marqo.tensor_search.on_start_script import on_start
 from marqo.tensor_search.telemetry import RequestMetricsStore, TelemetryMiddleware
