+++ src/marqo/tensor_search/models/api_models.py
@@ -9,7 +9,6 @@ from typing import Union, List, Dict, Optional
 import pydantic
 from pydantic import BaseModel, root_validator
 
-from marqo.core.models.interpolation_method import InterpolationMethod
 from marqo.core.models.marqo_index import MarqoIndex
 from marqo.tensor_search import validation
 from marqo.tensor_search.enums import SearchMethod
@@ -74,23 +73,6 @@ class SearchQuery(BaseMarqoModel):
     def get_context_tensor(self) -> Optional[List[SearchContextTensor]]:
         """Extract the tensor from the context, if provided"""
         return self.context.tensor if self.context is not None else None
-
-
-class RecommendQuery(BaseMarqoModel):
-    documents: Union[List[str], Dict[str, float]]
-    tensorFields: Optional[List[str]] = None
-    interpolationMethod: Optional[InterpolationMethod] = None
-    excludeInputDocuments: bool = True
-    limit: int = 10
-    offset: int = 0
-    efSearch: Optional[int] = None
-    approximate: Optional[bool] = None
-    searchableAttributes: Optional[List[str]] = None
-    showHighlights: bool = True
-    reRanker: str = None
-    filter: str = None
-    attributesToRetrieve: Union[None, List[str]] = None
-    scoreModifiers: Optional[ScoreModifier] = None
 
 
 class BulkSearchQueryEntity(SearchQuery):
